use specs2 for e2e test
we use it for twirp-runtime unit tests already,
so this makes things more consistent.
