+++ codegen/src/sbt-test/generator/e2e/build.sbt
@@ -8,4 +8,4 @@ scalacOptions ++= Seq(
   "-Xfatal-warnings"
 )
 
-libraryDependencies += "org.scalatest" %% "scalatest" % "3.0.5" % "test"
+libraryDependencies += "org.specs2" %% "specs2-core" % "4.3.6" % Test
